don't rebuild submission details on edit
If the webform doesn't have a details field, this works reasonably
well. But if it does have a Details field, the fields simply gets
all of this information appended to it every time the submission
is edited. With this change, the Details field only records the
initial submission, not any updates.